backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
gunicorn
python-multipart
supabase
openai
PyPDF2
mammoth
git+https://github.com/justdoitDT/textract@master#egg=textract
numpy
PyJWT>=2.8
httpx>=0.26,<0.29
email-validator
python-jose[cryptography]
requests